requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via django
bleach==6.2.0
    # via martor
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
distro==1.9.0
    # via openai
django==4.2
    # via
    #   shzleeblog (pyproject.toml)
    #   martor
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.5.0
    # via markdown
jiter==0.9.0
    # via openai
markdown==3.5.2
    # via martor
martor==1.6.45
    # via shzleeblog (pyproject.toml)
openai==1.66.5
    # via shzleeblog (pyproject.toml)
pillow==11.0.0
    # via shzleeblog (pyproject.toml)
psycopg2==2.9.10
    # via shzleeblog (pyproject.toml)
pydantic==2.10.6
    # via
    #   shzleeblog (pyproject.toml)
    #   openai
pydantic-core==2.27.2
    # via pydantic
requests==2.32.3
    # via martor
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.1
    # via django
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via martor
urllib3==2.2.3
    # via requests
uwsgi==2.0.28
    # via shzleeblog (pyproject.toml)
webencodings==0.5.1
    # via bleach
zipp==3.20.2
    # via importlib-metadata
